requirements.txt
python-dotenv
lightly
pillow
